requirements.txt
aiogram>=3.4.1
SQLAlchemy>=2.0.27
asyncpg>=0.29.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
aiohttp>=3.9.3
redis>=5.0.1
openpyxl>=3.1.2
pandas>=2.2.0
apscheduler>=3.10.4
